# server/lib/v2/api.yaml
openapi: '3.0.3'
info:
  version: 2.0.0
  title: Coin Wallet API V2
servers:
  - url: /api/v2
paths:
  /register:
    post:
      x-eov-operation-id: register
      x-eov-operation-handler: handlers
      description: Register new device
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: 'object'
              properties:
                walletId:
                  description: Wallet ID
                  $ref: '#/components/schemas/Hex64String'
                deviceId:
                  description: Device ID
                  $ref: '#/components/schemas/Hex64String'
                pinHash:
                  description: PIN Hash
                  $ref: '#/components/schemas/Hex64String'
              required:
                - walletId
                - deviceId
                - pinHash
      responses:
        '201':
          description: Registration Success
          content:
            'application/json':
              schema:
                allOf:
                  - $ref: '#/components/schemas/PublicToken'
                  - $ref: '#/components/schemas/PrivateToken'
        default:
          $ref: '#/components/responses/GeneralError'
...
  /details:
    get:
      x-eov-operation-id: getDetails
      x-eov-operation-handler: handlers
      description: Get user details
      parameters:
        - $ref: '#/components/parameters/deviceId'
      security:
        - deviceSignature: []
      responses:
        '200':
          description: User datails
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    description: User details
                    type: string
                required:
                  - data
        default:
          $ref: '#/components/responses/GeneralError'
    put:
      x-eov-operation-id: setDetails
      x-eov-operation-handler: handlers
      description: Update user details
      parameters:
        - $ref: '#/components/parameters/deviceId'
      security:
        - deviceSignature: []
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: 'object'
              properties:
                data:
                  description: User details
                  type: string
              required:
                - data
      responses:
        '200':
          description: Update user details result
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                required:
                  - data
        default:
          $ref: '#/components/responses/GeneralError'
  /username:
    put:
      x-eov-operation-id: setUsername
      x-eov-operation-handler: handlers
      description: Update username
      parameters:
        - $ref: '#/components/parameters/deviceId'
      security:
        - deviceSignature: []
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/Username'
      responses:
        '200':
          description: Update username result
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Username'
        default:
          $ref: '#/components/responses/GeneralError'
...
components:
  schemas:
    Hex64String:
      type: string
      pattern: "^[a-f0-9]{64}$"
    Hex128String:
      type: string
      pattern: "^[a-f0-9]{128}$"
    PublicToken:
      type: object
      properties:
        publicToken:
          $ref: '#/components/schemas/Hex128String'
      required:
        - publicToken
    PrivateToken:
      type: object
      properties:
        privateToken:
          $ref: '#/components/schemas/Hex128String'
      required:
        - privateToken
...
    Username:
      type: object
      properties:
        username:
          type: string
      required:
        - username
...
  parameters:
    deviceId:
      in: query
      name: id
      required: true
      description: Device ID
      schema:
        $ref: '#/components/schemas/Hex64String'
...
  responses:
...
    GeneralError:
      description: Unexpected error
      content:
        'application/json':
          schema:
            required:
              - code
              - message
            properties:
              code:
                type: integer
                format: int32
              message:
                type: string
  securitySchemes:
    deviceSignature:
      type: apiKey
      in: header
      name: Signature
    walletSignature:
      type: apiKey
      in: header
      name: Signature